snowplow_tracker v0.1.0 SnowplowTracker

Module is used to track events and send them to snowplow

Functions

track_ecommerce_transaction(event, tracker, module \\ Events.EcommerceTransaction)

track_ecommerce_transaction(
  SnowplowTracker.Events.EcommerceTransaction.t(),
  SnowplowTracker.Tracker.t(),
  module()
) :: {:ok, String.t()} | no_return()

This function is used to track a EcommerceTransaction event.

track_ecommerce_transaction_item(event, tracker, module \\ Events.EcommerceTransactionItem)

track_ecommerce_transaction_item(
  SnowplowTracker.Events.EcommerceTransactionItem.t(),
  SnowplowTracker.Tracker.t(),
  module()
) :: {:ok, String.t()} | no_return()

This function is used to track a EcommerceTransactionItem event.

track_page_view(event, tracker, module \\ Events.PageView)

track_page_view(
  SnowplowTracker.Events.PageView.t(),
  SnowplowTracker.Tracker.t(),
  module()
) :: {:ok, String.t()} | no_return()

This function is used to track a PageView event.

track_screen_view(event, tracker, module \\ Events.ScreenView)

track_screen_view(
  SnowplowTracker.Events.ScreenView.t(),
  SnowplowTracker.Tracker.t(),
  module()
) :: {:ok, String.t()} | no_return()

This function is used to track a ScreenView event.

track_self_describing_event(event, tracker, module \\ Events.SelfDescribing)

track_self_describing_event(
  SnowplowTracker.Events.SelfDescribing.t(),
  SnowplowTracker.Tracker.t(),
  module()
) :: {:ok, String.t()} | no_return()

This function is used to track a SelfDescribing event.

track_struct_event(event, tracker, module \\ Events.Structured)

track_struct_event(
  SnowplowTracker.Events.Structured.t(),
  SnowplowTracker.Tracker.t(),
  module()
) :: {:ok, String.t()} | no_return()

This function is used to track a Structured event.

track_timing(event, tracker, module \\ Events.Timing)

track_timing(
  SnowplowTracker.Events.Timing.t(),
  SnowplowTracker.Tracker.t(),
  module()
) :: {:ok, String.t()} | no_return()

This function is used to track a Timing event.