snowplow_tracker v0.1.0 SnowplowTracker.Constants

This module contains the constants used to construct the event payload and send them to the collector.

Examples

iex> SnowplowTracker.Constants.tracker_version()
"elixir-0.1.0"

Functions

app_id()

color_depth()

context()

context_encoded()

data()

domain_uid()

eid()

event()

event_ecomm()

event_ecomm_item()

event_page_view()

event_structured()

event_unstructured()

get_protocol_path()

ip_address()

language()

namespace()

network_uid()

page_refr()

page_title()

page_url()

platform()

post_content_type()

post_protocol_vendor()

post_protocol_version()

resolution()

schema()

schema_contexts()

schema_payload_data()

schema_screen_view()

schema_unstruct_event()

schema_user_timings()

se_action()

se_category()

se_label()

se_property()

se_value()

sent_timestamp()

sv_id()

sv_name()

t_version()

ti_item_category()

ti_item_currency()

ti_item_id()

ti_item_name()

ti_item_price()

ti_item_quantity()

ti_item_sku()

timestamp()

timezone()

tr_affiliation()

tr_city()

tr_country()

tr_currency()

tr_id()

tr_shipping()

tr_state()

tr_tax()

tr_total()

tracker_version()

true_timestamp()

uid()

unstructured()

unstructured_encoded()

useragent()

ut_category()

ut_label()

ut_timing()

ut_variable()

viewport()