snowplow_tracker v0.1.0 SnowplowTracker.Events.Helper

This module contains the implementations of function used to set the default values in the payload.

Functions

generate_timestamp(module \\ :os)

generate_timestamp(module()) :: Integer.t()

Generate unix timestamp in microseconds to identify time of each event.

generate_uuid()

generate_uuid() :: String.t()

Generate a v4 UUID string to uniquely identify an event

to_string(number)

to_string(any()) :: nil | String.t()

This function is used to convert a given number to a string. If the number is of type float, it is rounded off to 2 places and converted to a string.