snowplow_tracker v0.1.0 SnowplowTracker.Events.Structured

This module implements the Structured event.

Types

t()

t() :: %SnowplowTracker.Events.Structured{
  action: String.t(),
  category: String.t(),
  contexts: [SnowplowTracker.SelfDescribingJson.t()],
  event_id: String.t(),
  label: String.t(),
  property: String.t(),
  timestamp: integer(),
  true_timestamp: integer(),
  value: float()
}

Functions

get(event)

get(t()) :: SnowplowTracker.Payload.t()

This function should generate a map containing the key-value pairs for the event and return a payload object.

Callback implementation for SnowplowTracker.Events.Behaviour.get/1.

validate(event)

validate(t()) :: t() | no_return()

This function should take a struct() and validate the fields of the structure. If it is valid, it should return the object else it should raise an appropriate error.

Callback implementation for SnowplowTracker.Events.Behaviour.validate/1.