snowplow_tracker v0.1.0 SnowplowTracker.Payload

Represents the data structure used to store event information

Types

t()

t() :: %SnowplowTracker.Payload{pairs: map()}

Functions

add(payload, key, value)

add(t(), String.t() | Atom.t(), String.t() | Atom.t()) :: t()

This function is used to add a key-value map to the payload object

add_json(payload, map, key_encoded, key_not_encoded, encode)

add_json(t(), map(), String.t(), String.t(), boolean()) :: t()

This function is used convert a map to json, encode it as a base64 string depending on the encode flag and add it as a key-value pair to a payload object.

add_map(payload, map)

add_map(t(), map()) :: t()

This function is used to add a map containing one or many key-value pairs to a payload object.

get(payload)

get(t()) :: map()

This function is used to get the map containing key-value pairs stored in a payload object.

string(payload, encode)

string(t(), boolean()) :: t()

This function is used to convert a payload object to string and encode it to base64 depending on the encode flag.